+++ R/addpoly.default.r
@@ -92,8 +92,9 @@ transf, atransf, targs, efac, col, border, lty, fonts, cex, constarea=FALSE, ...
       preddist <- ddd$preddist
       if (!is.list(preddist))
          stop(mstyle$stop("Argument 'preddist' must be a list."))
-      dx  <- diff(preddist[[1]])[1]
-      cdf <- cumsum(preddist[[2]]) * dx
+      #dx  <- diff(preddist[[1]])[1]
+      #cdf <- cumsum(preddist[[2]]) * dx
+      cdf <- mapply(.trapezoid, preddist[[1]], preddist[[2]])
       cdf <- cdf / max(cdf)
       pi.lb <- preddist[[1]][which.min(abs(cdf - pi.level/2))]
       pi.ub <- preddist[[1]][which.min(abs(cdf - (1-pi.level/2)))]