+++ R/llplot.r
@@ -335,10 +335,9 @@ lty, lwd, col, level=99.99, refline=0, ...) {
    }
 
    if (scale) {
-      trapezoid <- function(x,y) sum(diff(x)*(y[-1]+y[-length(y)]))/2
       lls.sum <- rep(NA_real_, k)
       for (i in seq_len(k)) {
-         lls.sum[i] <- trapezoid(xs[!is.na(lls[i,])], lls[i,!is.na(lls[i,])])
+         lls.sum[i] <- .trapezoid(xs[!is.na(lls[i,])], lls[i,!is.na(lls[i,])])
          lls[i,] <- lls[i,] / lls.sum[i]
       }
    }